refactor #529: split the outcome reporting out of cookies.go
The 800-line ceiling caught cookies.go at 901 after the outcome work landed.
Split rather than allowlisted, per the operator's standing preference that
oversized files be refactored and the gates left to do their job.

The seam is a real one rather than a line-count convenience. Everything moved to
cookie_outcome.go is about naming WHY a read produced what it did and saying so
at a level a user sees: the outcome type, its String(), the reporting wrapper,
and hostForLog. The reading itself stays next to the cookie machinery it belongs
to.

901 -> 795 lines, no behaviour change.

Verification: make dod exit 0, read from the gate's own log.
